feat(desktop): move fns to specialized modules
